Update alembic_check.sh to use 'alembic upgrade head' instead of 'alembic stamp head'
