+++ alembic/versions/5128cc2fe488_remove_fk_in_document_able.py
@@ -1,0 +1,32 @@
+"""remove fk in document able
+
+Revision ID: 5128cc2fe488
+Revises: 73817166f499
+Create Date: 2024-12-04 17:14:10.177083
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '5128cc2fe488'
+down_revision: Union[str, None] = '73817166f499'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_constraint('documents_work_id_fkey', 'documents', type_='foreignkey')
+    op.drop_column('documents', 'work_id')
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('documents', sa.Column('work_id', sa.INTEGER(), autoincrement=False, nullable=True))
+    op.create_foreign_key('documents_work_id_fkey', 'documents', 'works', ['work_id'], ['id'])
+    # ### end Alembic commands ###
